pyproject.toml
[project]
name = "coffeebeans-voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.1",
    "fastapi>=0.119.0",
    "ffmpeg>=1.4",
    "ffmpeg-python>=0.2.0",
    "google-cloud-speech>=2.33.0",
    "google-cloud-texttospeech>=2.31.0",
    "groq>=0.32.0",
    "httpx>=0.28.1",
    "ipykernel>=7.0.1",
    "langgraph>=1.0.0",
    "openai>=2.5.0",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "twilio>=9.8.4",
    "uvicorn>=0.37.0",
    "websockets>=15.0.1",
]
